Move metric tests to seperate module
