+++ instrumentation-genai/opentelemetry-instrumentation-openai-v2/tests/cassettes/test_async_chat_completion_metrics.yaml
@@ -64,6 +64,7 @@ interactions:
               "finish_reason": "stop"
             }
           ],
+          "service_tier": "default",
           "usage": {
             "prompt_tokens": 12,
             "completion_tokens": 12,
